# api/internal/handlers/swagger.yaml
openapi: 3.0.3
info:
  title: StreamSpace API
  description: |
    StreamSpace is a platform-agnostic container streaming platform that delivers GUI applications to web browsers.

    ## Architecture
    - **Control Plane**: Centralized API with WebSocket Hub for agent coordination
    - **Agents**: Lightweight executors on target platforms (Kubernetes, Docker)
    - **VNC Proxy**: Secure, firewall-friendly VNC streaming through WebSocket

    ## Authentication
    - **Users**: JWT tokens (24-hour expiration) via `/api/v1/auth/login`
    - **Agents**: API keys via `/api/v1/agents/register`
    - **Webhooks**: HMAC signatures

    ## Rate Limiting
    Default: 60 requests/minute per IP (configurable via `RATE_LIMIT_REQUESTS_PER_MINUTE`)

    ## Common Response Formats
    All error responses follow:
    ```json
    {
      "error": "ErrorCode",
      "message": "Detailed error message"
    }
    ```
  version: 2.0.0-beta
  contact:
    name: StreamSpace Team
    url: https://github.com/streamspace-dev/streamspace
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8000
    description: Local development
  - url: https://api.streamspace.example.com
    description: Production
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT token obtained from /api/v1/auth/login
    apiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key for agent authentication
  schemas:
...
    Resources:
      type: object
      properties:
        memory:
          type: string
          example: "2Gi"
        cpu:
          type: string
          example: "1000m"
    Template:
      type: object
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
          description: Unique template identifier
        display_name:
          type: string
        description:
          type: string
        category:
          type: string
        app_type:
          type: string
        icon_url:
          type: string
        tags:
          type: array
          items:
            type: string
        default_resources:
          $ref: '#/components/schemas/Resources'
        manifest:
          type: object
          description: Platform-specific deployment manifest
        created_at:
          type: string
          format: date-time
...
paths:
...
  /api/v1/templates/{id}:
    get:
      summary: Get template
      operationId: getTemplate
      tags:
        - templates
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Template'

    patch:
      summary: Update template
      operationId: updateTemplate
      tags:
        - templates
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '200':
          description: Template updated

    delete:
      summary: Delete template
      operationId: deleteTemplate
      tags:
        - templates
      security:
        - bearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        '204':
          description: Template deleted

  # Catalog
...